+++ lib/exshome_automation/web/live/index.ex
@@ -3,8 +3,14 @@ defmodule ExshomeAutomation.Web.Live.Index do
   Main automation page.
   """
   alias ExshomeAutomation.Services.VariableRegistry
+  alias ExshomeAutomation.Web.Live.ShowVariableModal
 
   use ExshomeWeb.Live.AppPage,
     dependencies: [{VariableRegistry, :variables}],
     icon: "📃"
+
+  @impl LiveView
+  def handle_event("show_variable", %{"id" => id}, socket) do
+    {:noreply, open_modal(socket, ShowVariableModal, %{"variable_id" => id})}
+  end
 end